pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "p2"
version = "0.1.0"
description = "P2: Deep Reinforcement Learning for HUNL Texas Hold'em"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Your Name", email = "your.email@example.com" },
]
dependencies = [
  "torch>=2.8.0",
  "pyyaml>=5.4.0",
  "wandb>=0.15.0",
  "hydra-core>=1.3.0",
  "omegaconf>=2.3.0",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Games/Entertainment",
]

[dependency-groups]
dev = [
  "pre-commit>=3.0.0",
  "pytest>=7.0.0",
  "line-profiler>=4.0.0",
  "ruff>=0.14.11",
  "ty>=0.0.11",
]

[project.urls]
Homepage = "https://github.com/phulin/poker2"

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  { index = "pytorch-cu130", marker = "sys_platform == 'linux'" },
]
torchvision = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  { index = "pytorch-cu130", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py311", "py312", "py313"]

[tool.isort]
profile = "black"
line_length = 88
src_paths = ["src", "tests", "debugging"]
known_first_party = ["p2"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning:wandb.*",
    "ignore::DeprecationWarning:wandb.analytics.sentry.*",
]
